version 0.0.2

+++ Source/PhotinoAPIWindow.cs
@@ -11,7 +11,10 @@ namespace PhotinoNET {
     public class PhotinoAPIWindow: PhotinoWindow {
         public object JS_API;
         private List<String> excludeNames = new List<String>() { "GetType", "ToString", "Equals", "GetHashCode" };
+        private String apiJsFile = null;
         private String tempHtmlFile = null;
+        private String tempIconFile = null;
+        private Boolean removeTempFile = false;
 
         private List<String> apiMethodNames {
             get => this.JS_API.GetType().GetMethods().Where(
@@ -31,8 +34,18 @@ namespace PhotinoNET {
                 result["state"] = "success";
                 result["methods"] = this.apiMethodNames;
 
-                if (this.tempHtmlFile != null) {
-                    File.Delete(this.tempHtmlFile);
+                if (this.removeTempFile || !this.DevToolsEnabled) {
+                    if (this.apiJsFile != null) {
+                        File.Delete(this.apiJsFile);
+                    }
+
+                    if (this.tempHtmlFile != null) {
+                        File.Delete(this.tempHtmlFile);
+                    }
+
+                    if (this.tempIconFile != null) {
+                        File.Delete(this.tempIconFile);
+                    }
                 }
             }
             else if (this.apiMethodNames.Contains(data["job"])) {
@@ -64,6 +77,7 @@ namespace PhotinoNET {
 
         public PhotinoAPIWindow RegisterAPI(object jsapi) {
             this.JS_API = jsapi;
+            this.Log($".RegisterAPI({jsapi})");
             return this;
         }
         public new PhotinoAPIWindow Center() {
@@ -79,29 +93,36 @@ namespace PhotinoNET {
             return this;
         }
         public new PhotinoAPIWindow Load(String path) {
-            String jsScript;
-            var assembly = System.Reflection.Assembly.GetExecutingAssembly();
-            using (var stream = assembly.GetManifestResourceStream($"{assembly.GetName().Name}.photino.net.api.js")) {
-                using (var reader = new StreamReader(stream)) {
-                    jsScript = reader.ReadToEnd();
+            throw new NotImplementedException();
+        }
+        public PhotinoAPIWindow LoadFile(String path, Boolean includeApiScript = true) {
+            if (includeApiScript) {
+                // String jsScript;
+                this.apiJsFile = Path.Join(Path.GetDirectoryName(path), ".photino.net.api.min.js");
+                var assembly = System.Reflection.Assembly.GetExecutingAssembly();
+                using (var stream = assembly.GetManifestResourceStream($"{assembly.GetName().Name}.photino.net.api.min.js")) {
+                    using (var reader = new StreamReader(stream)) {
+                        File.WriteAllText(this.apiJsFile, reader.ReadToEnd());
+                    }
                 }
-            }
 
-            var html = File.ReadAllText(path, encoding: System.Text.Encoding.UTF8);
-            html = html.Replace("<head>", $@"
-<head>
-    <script>
-{jsScript}
-    </script>
+                var html = File.ReadAllText(path, encoding: System.Text.Encoding.UTF8);
+                html = html.Replace("<head>", $@"
+    <head>
+        <script type='text/javascript' src='{this.apiJsFile}'></script>
 ");
 
-            this.tempHtmlFile = Path.Join(Path.GetDirectoryName(path), "." + Path.GetFileName(path));
-            File.WriteAllText(this.tempHtmlFile, html);
-            base.Load(this.tempHtmlFile);
+                this.tempHtmlFile = Path.Join(Path.GetDirectoryName(path), "." + Path.GetFileName(path));
+                File.WriteAllText(this.tempHtmlFile, html);
+                base.Load(this.tempHtmlFile);
+            }
+            else {
+                base.Load(path);
+            }
 
             return this;
         }
-        public new PhotinoAPIWindow Load(Uri uri) {
+        public PhotinoAPIWindow LoadUri(Uri uri) {
             base.Load(uri);
             return this;
         }
@@ -201,6 +222,17 @@ namespace PhotinoNET {
             base.SetIconFile(iconFile);
             return this;
         }
+        public PhotinoAPIWindow SetIconFromResource(String resourceName) {
+            var assembly = System.Reflection.Assembly.GetCallingAssembly();
+            this.tempIconFile = Path.Join(Path.GetDirectoryName(assembly.Location), resourceName);
+            using (var stream = assembly.GetManifestResourceStream(resourceName)) {
+                using (var writeStream = File.Create(this.tempIconFile)) {
+                    stream.CopyTo(writeStream);
+                }
+            }
+
+            return this.SetIconFile(this.tempIconFile);
+        }
         public new PhotinoAPIWindow SetLeft(Int32 left) {
             base.SetLeft(left);
             return this;
@@ -210,6 +242,7 @@ namespace PhotinoNET {
             return this;
         }
         public new PhotinoAPIWindow SetLogVerbosity(Int32 verbosity) {
+            base.LogVerbosity = verbosity;
             base.SetLogVerbosity(verbosity);
             return this;
         }
@@ -265,9 +298,18 @@ namespace PhotinoNET {
             base.SetZoom(zoom);
             return this;
         }
+        public PhotinoAPIWindow SetRemoveTempFile(Boolean removeTempFile) {
+            this.removeTempFile = removeTempFile;
+            return this;
+        }
         public new PhotinoAPIWindow Win32SetWebView2Path(String data) {
             base.Win32SetWebView2Path(data);
             return this;
+        }
+
+        private void Log(String message) {
+            if (this.LogVerbosity < 1) return;
+            Console.WriteLine($"Photino.NET: \"{this.Title ?? "PhotinoWindow"}\"{message}");
         }
     }
 }
